Change the behavior of pointer to object to double check that the pointer is a real pyobject.
This prevents many many different kinds of bad things from happening.
For example, if a non-pyobject is passed such as a callback
or userdata and the python code tries to print the last argument
python segfaults.   This is a reasonable safety measure.  There
may be a better way to check that its a valid pyobject but this works
reasonably well with high probability for now.

https://bugzilla.gnome.org/show_bug.cgi?id=615980